@@ -1,7 +1,7 @@
 {
     "name": "ngx-material-timepicker",
     "description": "Handy material design timepicker for angular",
-    "version": "1.1.2",
+    "version": "2.1.2",
     "license": "MIT",
     "author": "Vitalii Boiko <boyko330@gmail.com>",
     "keywords": [
@@ -43,24 +43,24 @@
         "@angular/core": ">=4.0.0",
         "@angular/forms": ">=4.0.0",
         "@angular/platform-browser": ">=4.0.0",
-        "rxjs": "^5.5.7"
+        "rxjs": ">=6.0.0"
     },
     "dependencies": {
         "moment": "^2.21.0"
     },
     "devDependencies": {
-        "@angular/animations": "^5.1.0",
-        "@angular/cli": "~1.7.3",
-        "@angular/common": "^5.1.0",
-        "@angular/compiler": "^5.1.0",
-        "@angular/compiler-cli": "^5.1.0",
-        "@angular/core": "^5.1.0",
-        "@angular/forms": "^5.1.0",
-        "@angular/http": "^5.1.0",
-        "@angular/language-service": "^5.1.0",
-        "@angular/platform-browser": "^5.1.0",
-        "@angular/platform-browser-dynamic": "^5.1.0",
-        "@angular/router": "^5.1.0",
+        "@angular/animations": "^6.0.3",
+        "@angular/cli": "^6.0.5",
+        "@angular/common": "^6.0.3",
+        "@angular/compiler": "^6.0.3",
+        "@angular/compiler-cli": "^6.0.3",
+        "@angular/core": "^6.0.3",
+        "@angular/forms": "^6.0.3",
+        "@angular/http": "^6.0.3",
+        "@angular/language-service": "^6.0.3",
+        "@angular/platform-browser": "^6.0.3",
+        "@angular/platform-browser-dynamic": "^6.0.3",
+        "@angular/router": "^6.0.3",
         "@types/jasmine": "~2.8.3",
         "@types/jasminewd2": "~2.0.2",
         "@types/node": "~6.0.60",
@@ -74,14 +74,13 @@
         "karma-jasmine": "~1.1.0",
         "karma-jasmine-html-reporter": "^0.2.2",
         "ng-packagr": "^2.2.0",
-        "ngx-material-timepicker": "^1.1.1",
         "protractor": "~5.1.2",
-        "rxjs": "^5.5.7",
-        "semantic-release": "^15.4.1",
+        "rxjs": "^6.2.0",
         "travis-deploy-once": "^5.0.0",
         "ts-node": "~4.1.0",
         "tslint": "~5.9.1",
-        "typescript": "2.4.2",
-        "zone.js": "^0.8.19"
+        "typescript": "2.7.2",
+        "zone.js": "^0.8.19",
+        "@angular-devkit/build-angular": "~0.6.5"
     }
 }
